+++ testutil/keeper/model.go
@@ -42,6 +42,7 @@ func ModelKeeper(t testing.TB) (*keeper.Keeper, sdk.Context) {
 		nil,
 		nil,
 		nil,
+		nil,
 		cdc,
 		storeKey,
 		memStoreKey,
