+++ testutil/keeper/node.go
@@ -39,6 +39,8 @@ func NodeKeeper(t testing.TB) (*keeper.Keeper, sdk.Context) {
 	k := keeper.NewKeeper(
 		nil,
 		nil,
+		nil,
+		nil,
 		cdc,
 		storeKey,
 		memStoreKey,
