+++ testutil/keeper/order.go
@@ -40,7 +40,6 @@ func OrderKeeper(t testing.TB) (*keeper.Keeper, sdk.Context) {
 		nil,
 		nil,
 		nil,
-		nil,
 		cdc,
 		storeKey,
 		memStoreKey,
