+++ testutil/keeper/sao.go
@@ -44,6 +44,7 @@ func SaoKeeper(t testing.TB) (*keeper.Keeper, sdk.Context) {
 		nil,
 		nil,
 		nil,
+		nil,
 		cdc,
 		storeKey,
 		memStoreKey,
