+++ x/model/client/cli/query_expired_order.go
@@ -3,7 +3,7 @@ package cli
 import (
 	"context"
 
-	"github.com/SaoNetwork/sao/x/order/types"
+	"github.com/SaoNetwork/sao/x/model/types"
 	"github.com/cosmos/cosmos-sdk/client"
 	"github.com/cosmos/cosmos-sdk/client/flags"
 	"github.com/spf13/cast"
