+++ x/model/keeper/grpc_query_expired_order.go
@@ -2,8 +2,8 @@ package keeper
 
 import (
 	"context"
+	"github.com/SaoNetwork/sao/x/model/types"
 
-	"github.com/SaoNetwork/sao/x/order/types"
 	"github.com/cosmos/cosmos-sdk/store/prefix"
 	sdk "github.com/cosmos/cosmos-sdk/types"
 	"github.com/cosmos/cosmos-sdk/types/query"
