+++ x/model/types/expected_keepers.go
@@ -29,6 +29,7 @@ type OrderKeeper interface {
 	GetOrder(ctx sdk.Context, orderId uint64) (ordertypes.Order, bool)
 	RemoveOrder(ctx sdk.Context, id uint64)
 	TerminateOrder(ctx sdk.Context, orderId uint64, refundCoin sdk.Coin) error
+	RefundOrder(ctx sdk.Context, orderId uint64) error
 }
 
 // SaoKeeper
