+++ x/order/genesis_test.go
@@ -14,14 +14,6 @@ func TestGenesis(t *testing.T) {
 	genesisState := types.GenesisState{
 		Params: types.DefaultParams(),
 
-		ExpiredOrderList: []types.ExpiredOrder{
-			{
-				Height: 0,
-			},
-			{
-				Height: 1,
-			},
-		},
 		// this line is used by starport scaffolding # genesis/test/state
 	}
 
@@ -33,6 +25,5 @@ func TestGenesis(t *testing.T) {
 	nullify.Fill(&genesisState)
 	nullify.Fill(got)
 
-	require.ElementsMatch(t, genesisState.ExpiredOrderList, got.ExpiredOrderList)
 	// this line is used by starport scaffolding # genesis/test/assert
 }
