+++ x/order/types/genesis.go
@@ -1,7 +1,7 @@
 package types
 
 import (
-	fmt "fmt"
+	"fmt"
 )
 
 // this line is used by starport scaffolding # genesis/types/import
@@ -12,7 +12,6 @@ const DefaultIndex uint64 = 1
 // DefaultGenesis returns the default genesis state
 func DefaultGenesis() *GenesisState {
 	return &GenesisState{
-		ExpiredOrderList: []ExpiredOrder{},
 		// this line is used by starport scaffolding # genesis/types/default
 		OrderList: []Order{},
 		ShardList: []Shard{},
@@ -47,16 +46,6 @@ func (gs GenesisState) Validate() error {
 			return fmt.Errorf("shard id should be lower or equal than the last id")
 		}
 		shardIdMap[elem.Id] = true
-	}
-	// Check for duplicated index in expiredOrder
-	expiredOrderIndexMap := make(map[string]struct{})
-
-	for _, elem := range gs.ExpiredOrderList {
-		index := string(ExpiredOrderKey(elem.Height))
-		if _, ok := expiredOrderIndexMap[index]; ok {
-			return fmt.Errorf("duplicated index for expiredOrder")
-		}
-		expiredOrderIndexMap[index] = struct{}{}
 	}
 	// this line is used by starport scaffolding # genesis/types/validate
