+++ x/order/types/genesis_test.go
@@ -19,34 +19,11 @@ func TestGenesisState_Validate(t *testing.T) {
 			valid:    true,
 		},
 		{
-			desc: "valid genesis state",
+			desc:     "valid genesis state",
 			genState: &types.GenesisState{
-
-				ExpiredOrderList: []types.ExpiredOrder{
-					{
-						Height: 0,
-					},
-					{
-						Height: 1,
-					},
-				},
 				// this line is used by starport scaffolding # types/genesis/validField
 			},
 			valid: true,
-		},
-		{
-			desc: "duplicated expiredOrder",
-			genState: &types.GenesisState{
-				ExpiredOrderList: []types.ExpiredOrder{
-					{
-						Height: 0,
-					},
-					{
-						Height: 0,
-					},
-				},
-			},
-			valid: false,
 		},
 		// this line is used by starport scaffolding # types/genesis/testcase
 	} {
